STORK-A/scripts/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.5=py37h261ae71_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py37_0
dbus=1.13.18=hb2f20db_0
efficientnet-pytorch=0.7.0=pyh9f0ad1d_0
expat=2.3.0=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=py37_1
glib=2.68.1=h36276a3_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
intel-openmp=2021.2.0=h06a4308_610
joblib=0.17.0=py_0
jpeg=9b=habf39ab_1
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=hff7bd54_1
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.4=py37h2531618_0
pcre=8.44=he6710b0_0
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
psutil=5.8.0=py37h27cfd23_1
pycparser=2.20=py_2
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.0.130_cudnn7.6.3_0
pytz=2020.1=py_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.6.2=py37h91f5cce_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sqlite=3.35.4=hdfb4753_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
time=1.8=h516909a_0
tk=8.6.10=hbc83047_0
torchvision=0.5.0=py37_cu100
tornado=6.1=py37h27cfd23_0
typing_extensions=3.7.4.3=pyha847dfd_0
tzdata=2020f=h52ac0ba_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
